USB: Allow USB HID autosuspend for HID only devices
Enable autosuspend for common USB HID only devices, such as keyboards and
mice.  To improve user experience, add logic to track display status to
control USB bus suspend.  If bus suspend/resume iterations occur, users
will experience some latency/overhead from the time the mouse/keyboard input
is received.  The sequence will be as follows:

1. USB HAL identifies a new USB device.
2. Check if the device contains USB HID interfaces only.
3. If so, add to a string list containing all USB HID device paths.
4. Based on backlight uevents, adjust the auto-suspend timer accordingly.
   - 60s - Screen on
   - 3s  - Screen off
5. When device is removed, device path is removed from the string list.

The reason for applying this on devices that only expose USB HID
interfaces, is due to the existence of USB composite devices that have
other interfaces associated, such as USB headsets.  In this case, the USB
HID interface is the volume control, and another interface contains the
audio path.  For these scenarios, we don't want to employ the USB suspend
logic based on the display status, as suspend/resume should be based on the
audio stream status.

Change-Id: Ief1fc5ea97bec72f7c9a910e93dcf0133d5415c0